+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="v0.61.4"></a>
+## [v0.61.4] - 2021-07-28
+### Bug Fixes
+- update error handling code for alert policies
+
+<a name="v0.61.3"></a>
+## [v0.61.3] - 2021-07-28
+### Bug Fixes
+- retire usages of deprecated error schema
+
 <a name="v0.61.2"></a>
 ## [v0.61.2] - 2021-07-22
 ### Bug Fixes
@@ -1092,7 +1102,9 @@
 - extract paging implementation
 - rename packages for clarity, promote Config to the public package
 
-[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.61.2...HEAD
+[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.61.4...HEAD
+[v0.61.4]: https://github.com/newrelic/newrelic-client-go/compare/v0.61.3...v0.61.4
+[v0.61.3]: https://github.com/newrelic/newrelic-client-go/compare/v0.61.2...v0.61.3
 [v0.61.2]: https://github.com/newrelic/newrelic-client-go/compare/v0.61.1...v0.61.2
 [v0.61.1]: https://github.com/newrelic/newrelic-client-go/compare/v0.61.0...v0.61.1
 [v0.61.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.60.2...v0.61.0